filetobytesを移動
