+++ integration_test/tests/TestIOMulti.go
@@ -3,7 +3,6 @@ package tests
 import (
 	"integration_test/response"
 	"integration_test/tester"
-	"integration_test/tests/utils"
 	"net/http"
 )
 
@@ -24,7 +23,7 @@ var testIOMulti = testCatergory{
 					},
 					ExpectStatusCode: http.StatusOK,
 					ExpectHeader:     nil,
-					ExpectBody:       utils.FileToBytes("../html/index.html"),
+					ExpectBody:       fileToBytes("../html/index.html"),
 				})
 
 				clientB := tester.NewClient(&tester.Client{
